+++ skbio/io/format/blast6.py
@@ -1,10 +1,10 @@
 """
 BLAST+6 format (:mod:`skbio.io.format.blast6`)
-=========================================================================
+==============================================
 
 .. currentmodule:: skbio.io.format.blast6
 
-The BLAST+6 format (``blast6``) stores strings, integers, and floats relating
+The BLAST+6 format (``blast+6``) stores strings, integers, and floats relating
 to the similarities between multiple DNA sequences. These values are stored in
 a simple tabular format with no labels on individual columns. The values are
 separated by the tab character.
@@ -12,14 +12,15 @@ separated by the tab character.
 An example BLAST+6-formatted file comparing two protein sequences, taken
 from [1]_::
 
-    moaC    gi|15800534|ref|NP_286546.1|    100.00  161 0   0   1   161 1
-    161 3e-114  330
-    moaC    gi|170768970|ref|ZP_02903423.1| 99.38   161 1   0   1   161 1
-    161 9e-114  329
+    moaC<tab>gi|15800534|ref|NP_286546.1|<tab>100.00<tab>161<tab>0<tab>0<tab>1\
+    <tab>161<tab>1<tab>161<tab>3e-114<tab>330
+    moaC<tab>gi|170768970|ref|ZP_02903423.1|<tab>99.38<tab>161<tab>1<tab>0\
+    <tab>1<tab>161<tab>1<tab>161<tab>9e-114<tab>329
 
 Format Support
 --------------
 **Has Sniffer: No**
+**State: Experimental as of 0.4.0-dev.**
 
 +------+------+---------------------------------------------------------------+
 |Reader|Writer|                          Object Class                         |
@@ -31,7 +32,8 @@ Format Specification
 --------------------
 BLAST+6 format is a tabular text-based format produced by BLAST+ format 6,
 containing data related to passed sequences, tab-separated, with no column
-labels or headers of any kind.
+labels or headers of any kind. This format includes both BLAST+ output format
+6 and BLAST output format 8.
 
 When blasting sequences using Ubuntu's command line, the user may specify what
 data is returned.
@@ -54,9 +56,8 @@ For example::
     -max_target_seqs 2 -outfmt 6
     $ query1	subject2	100.00	8	0	0	1	8	3	10	9e-05	16.9
 
-If the user does not pass custom values, the blastp command will use the
-default columns, which are qseqid, sseqid, pident, length, mismatch, gapopen,
-qstart, qend, sstart, send, evalue, and bitscore.
+If the user does not pass custom values, the blastp command will use its
+default column values.
 
 Types of Data
 ^^^^^^^^^^^^^
@@ -75,7 +76,7 @@ Types of Data
 +-----------+------------------------------------+-----+
 |sseqid     |Subject Seq-id                      |str  |
 +-----------+------------------------------------+-----+
-|sallseqid  |All subject Seq-id(s), separated by\|str  |
+|sallseqid  |All subject Seq-id(s), separated by |str  |
 |           |a ';'                               |     |
 +-----------+------------------------------------+-----+
 |sgi        |Subject GI                          |int  |
@@ -125,7 +126,7 @@ Types of Data
 |ppos       |Percentage of positive-scoring matc\|float|
 |           |hes                                 |     |
 +-----------+------------------------------------+-----+
-|frames     |Query and subject frames separated \|str  |
+|frames     |Query and subject frames separated  |str  |
 |           |by a '/'                            |     |
 +-----------+------------------------------------+-----+
 |qframe     |Query frame                         |int  |
@@ -137,25 +138,25 @@ Types of Data
 |staxids    |Unique Subject Taxonomy ID(s), sepa\|str  |
 |           |rated by a ';' (in numerical order) |     |
 +-----------+------------------------------------+-----+
-|sscinames  |Unique Subject Scientific Name(s), \|str  |
+|sscinames  |Unique Subject Scientific Name(s),  |str  |
 |           |separated by a ';'                  |     |
 +-----------+------------------------------------+-----+
 |scomnames  |Unique Subject Common Name(s), sepa\|str  |
 |           |rated by a ';'                      |     |
 +-----------+------------------------------------+-----+
 |sblastnames|unique Subject Blast Name(s), separ\|str  |
-|           |ated by a ';' (in alphabetical\     |     |
+|           |ated by a ';' (in alphabetical      |     |
 |           |order)                              |     |
 +-----------+------------------------------------+-----+
 |sskingdoms |unique Subject Super Kingdom(s), se\|str  |
-|           |parated by a ';' (in alphabetical\  |     |
+|           |parated by a ';' (in alphabetical   |     |
 |           |order)                              |     |
 +-----------+------------------------------------+-----+
 |stitle     |Subject Title                       |str  |
 +-----------+------------------------------------+-----+
 |sstrand    |Subject Strand                      |str  |
 +-----------+------------------------------------+-----+
-|salltitles |All Subject Title(s), separated by \|str  |
+|salltitles |All Subject Title(s), separated by  |str  |
 |           |a '<>'                              |     |
 +-----------+------------------------------------+-----+
 |qcovs      |Query Coverage Per Subject          |int  |
@@ -163,57 +164,72 @@ Types of Data
 |qcovhsp    |Query Coverage Per HSP              |int  |
 +-----------+------------------------------------+-----+
 
+When a blastp data type returns NA values, blastp displays it as
+``'N/A'``. When read into a pd.DataFrame, these values are converted
+into ``np.nan``.
+
 List of data and descriptions taken from BLAST Command Line Applications User
 Manual [2]_.
 
+Format Parameters
+-----------------
+The following format parameters are available in ``blast+6`` format only:
+
+-``default_columns``: ``False`` by default. If ``True``, will set the column
+ labels to the default BLAST+ column labels, which are qseqid, sseqid, pident,
+ length, mismatch, gapopen, qstart, qend, sstart, send, evalue, and bitscore.
+
+-``columns``: ``None`` by default. Accepts a list of labels which will be
+ assigned to the returned pd.Dataframe.
+
 Examples
 --------
-An example of using _blast6_to_data_frame with default column labels.
+An example of reading in ``blast+6`` format with default column labels.
 
 >>> from io import StringIO
->>> from skbio import read
+>>> import skbio.io as io
 >>> import pandas as pd
 >>> fs = '\\n'.join([
-...     'query1\\tsubject1\\t30.0\\t3\\t5\\t1\\t1\\t3\\t8\\t7\\t0.030\\t13.2',
-...     'query2\\tsubject2\\t40.0\\t5\\t4\\t2\\t1\\t6\\t9\\t0\\t0.045\\t15.0'
+...     'moaC\\tgi|15800534|ref|NP_286546.1|\\t100.00\\t161\\t0\\t0\\t1\\t161\
+\\t1\\t161\\t3e-114\\t330',
+...     'moaC\\tgi|170768970|ref|ZP_02903423.1|\\t99.38\\t161\\t1\\t0\\t1\\t\
+161\\t1\\t161\\t9e-114\\t329'
 ... ])
 >>> fh = StringIO(fs)
 
-Load BLAST+6 data into a DataFrame and specify that default columns should be
-used.
+Load BLAST+6 data into a pd.Dataframe and specify that default columns should
+be used.
 
->>> df = read(fh, format="blast+6", into=pd.DataFrame, default_columns=True)
+>>> df = io.read(fh, format="blast+6", into=pd.DataFrame, default_columns=True)
 >>> df # doctest: +NORMALIZE_WHITESPACE
-   qseqid    sseqid  pident  length  mismatch  gapopen  qstart  qend  sstart \\
-0  query1  subject1      30       3         5        1       1     3       8
-1  query2  subject2      40       5         4        2       1     6       9
+  qseqid                           sseqid  pident  length  mismatch  gapopen \\
+0   moaC     gi|15800534|ref|NP_286546.1|  100.00     161         0        0
+1   moaC  gi|170768970|ref|ZP_02903423.1|   99.38     161         1        0
 <BLANKLINE>
-   send  evalue  bitscore
-0     7   0.030      13.2
-1     0   0.045      15.0
+   qstart  qend  sstart  send         evalue  bitscore
+0       1   161       1   161  3.000000e-114       330
+1       1   161       1   161  9.000000e-114       329
 
-An example of using _blast6_to_data_frame with custom column labels.
+An example of reading in ``blast+6`` format with custom column labels.
 
 >>> from io import StringIO
->>> from skbio import read
+>>> import skbio.io as io
 >>> import pandas as pd
 >>> fs = '\\n'.join([
-...     'subject2\\tquery1\\t100\\tN/A\\t0\\tsubject2\\t32\\t100.0',
-...     'subject1\\tquery2\\t70\\tN/A\\t0\\tsubject1\\t19\\t70.0',
-...     'subject2\\tquery1\\t100\\tN/A\\t0\\tsubject2\\t18\\t50.0'
+...     'moaC\\t100.00\\t0\\t161\\t0\\t161\\t330\\t1',
+...     'moaC\\t99.38\\t1\\t161\\t0\\t161\\t329\\t1'
 ... ])
 >>> fh = StringIO(fs)
 
-Load BLAST+6 data into a DataFrame and specify your column types.
+Load BLAST+6 data into a pd.Dataframe and specify your column types.
 
->>> df = read(fh, format="blast+6", into=pd.DataFrame,
-...           columns=['sacc', 'qaccver', 'qcovs', 'sblastnames', 'gapopen',
-...                    'sallacc', 'score', 'ppos'])
+>>> df = io.read(fh, format="blast+6", into=pd.DataFrame,
+...           columns=['qseqid', 'pident', 'mismatch', 'length', 'gapopen',
+...                    'qend', 'bitscore', 'sstart'])
 >>> df # doctest: +NORMALIZE_WHITESPACE
-       sacc qaccver  qcovs  sblastnames  gapopen   sallacc  score  ppos
-0  subject2  query1    100          NaN        0  subject2     32   100
-1  subject1  query2     70          NaN        0  subject1     19    70
-2  subject2  query1    100          NaN        0  subject2     18    50
+  qseqid  pident  mismatch  length  gapopen  qend  bitscore  sstart
+0   moaC  100.00         0     161        0   161       330       1
+1   moaC   99.38         1     161        0   161       329       1
 
 References
 ----------
@@ -232,9 +248,11 @@ and-csv.html
 
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
-from skbio.io import create_format
+
 import pandas as pd
 import numpy as np
+
+from skbio.io import create_format, BLAST6FormatError
 
 blast6 = create_format('blast+6')
 
@@ -256,32 +274,38 @@ _default_columns = ['qseqid', 'sseqid', 'pident', 'length', 'mismatch',
                     'evalue', 'bitscore']
 
 
-@blast6.reader(pd.DataFrame)
+@blast6.reader(pd.DataFrame, monkey_patch=False)
 def _blast6_to_data_frame(fh, columns=None, default_columns=False):
-    if default_columns is True and columns is not None:
-        raise ValueError("You cannot use the default columns and also create"
-                         " your own columns")
-    if default_columns is False and columns is None:
-        raise ValueError("You must specify whether default columns or custom"
-                         " columns are being used.")
-    if columns is not None:
-        names = columns
-    else:
-        names = _default_columns
+    if default_columns and columns is not None:
+        raise ValueError("`columns` and `default_columns` cannot both be"
+                         " provided.")
+    if not default_columns and columns is None:
+        raise ValueError("Either `columns` or `default_columns` must be"
+                         " provided.")
+    if default_columns:
+        columns = _default_columns
 
     df = pd.read_csv(fh, na_values='N/A', sep='\t', header=None)
 
-    if len(list(df.columns.values)) != len(names):
-        raise ValueError("The amount of data you have supplied does not match"
-                         " the number of columns.")
-
-    try:
-        for i in range(len(names)):
-            df[i] = df[i].astype(_possible_columns[names[i]])
-            df.replace('nan', np.nan, inplace=True)
-        df.columns = names
-    except ValueError as e:
-        raise ValueError('Values have been assigned to incorrect columns. '
-                         'Original Pandas error message: '
-                         '"%s"' % e)
+    if len(df.columns) != len(columns):
+        raise BLAST6FormatError("The specified number of columns: %d does not"
+                                " match the number of columns in the file: "
+                                "%d." % (len(columns), len(df.columns)))
+    df.columns = columns
+    for column in df:
+        if column not in _possible_columns:
+            possible_columns = []
+            for key in _possible_columns:
+                possible_columns.append(key)
+            raise ValueError("Your column name: %s is not valid."
+                             "The valid column names are:\n%s"
+                             "" % (column, possible_columns))
+        df.replace('nan', np.nan, inplace=True)
+        try:
+            df[column] = df[column].astype(_possible_columns[column])
+        except ValueError as e:
+            raise BLAST6FormatError('Could not convert column %r into dtype'
+                                    '%r. Original Pandas error message: '
+                                    '"%s"' % (column,
+                                              _possible_columns[column], e))
     return df